Product settings update completed

+++ app/Http/Controllers/Api/ProductTypeController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\ProductTypeResource;
+use App\Models\Product;
 use App\Models\ProductType;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
@@ -30,6 +31,15 @@ class ProductTypeController extends Controller
      */
     public function store(Request $request): JsonResponse
     {
+        $name = $request->name;
+
+        // Check if the color already exists
+        $existingType = ProductType::where('name', $name)->first();
+        if ($existingType) {
+            return response()->json([
+                "message" => "Product type already exists."
+            ], 422);
+        }
         $productType = new ProductType();
         $productType->name = $request->name;
         $productType->save();
@@ -58,9 +68,35 @@ class ProductTypeController extends Controller
      * @param  \App\Models\ProductType  $productType
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, ProductType $productType)
+    public function update(Request $request, $id): JsonResponse
     {
-        //
+        $productType = ProductType::find($id);
+
+        if (!$productType) {
+            return response()->json([
+                "message" => "Product type not found."
+            ], 404);
+        }
+
+        // Update the product color record
+        $productType->name = $request->input('name');
+        try {
+            $productType->save();
+        } catch (\Illuminate\Database\QueryException $e) {
+            // Handle foreign key constraint violation error
+            if ($e->errorInfo[1] == 1451) {
+                return response()->json([
+                    "message" => "Cannot update product type."
+                ], 400);
+            } else {
+                throw $e;
+            }
+        }
+
+        return response()->json([
+            "message" => "Product type updated successfully.",
+            "data" => $productType
+        ]);
     }
 
     /**
@@ -77,6 +113,15 @@ class ProductTypeController extends Controller
             return response()->json([
                 "message" => "Product fabric not found."
             ], 404);
+        }
+
+        $referencingProducts = Product::where('type_id', $id)->get();
+
+        if (count($referencingProducts) > 0) {
+            // If there are referencing records, return an error message
+            return response()->json([
+                "message" => "Cannot delete product type because this is already in use."
+            ], 400);
         }
 
         $productType->delete();
